services/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
requests>=2.31.0
boto3>=1.35.0
pytest>=8.0.0
faiss-cpu>=1.8.0
numpy>=1.24.0
